+++ CHANGELOG.md
@@ -1,9 +1,19 @@
+## 0.2.0
+
+- add select widget
+
 ## 0.1.9
-- add single rounded 
+
+- add single rounded
+
 ## 0.1.8
+
 - fix boder property error
+
 ## 0.1.7
+
 - fix img flex1 bug
+
 ## 0.1.6
 
 - add example